backend/pyproject.toml
[project]
name = "cartoncaps-chat-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Austin Redenbaugh",email = "43683341+AustinRedenbaugh@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sse-starlette (>=2.3.6,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
